+++ .eslintrc.json
@@ -42,10 +42,6 @@
                 "allowTemplateLiterals": true
             }
         ],
-        "semi": [
-            "error",
-            "never"
-        ],
         "brace-style": [
             1,
             "1tbs",
@@ -59,29 +55,11 @@
             "never"
         ],
         "prefer-spread": 0,
-        "object-curly-spacing": [
-            "warn",
-            "always"
-        ],
         "array-bracket-spacing": [
             "warn",
             "never"
         ],
-        "comma-spacing": [
-            "error",
-            {
-                "before": false,
-                "after": true
-            }
-        ],
         "keyword-spacing": 1,
-        "key-spacing": [
-            "warn",
-            {
-                "beforeColon": false,
-                "afterColon": true
-            }
-        ],
         "arrow-body-style": 0,
         "no-multi-spaces": 1,
         "comma-dangle": 0,
