+++ tests/throwMsgHelper.d.ts
@@ -1,0 +1,1 @@
+export declare function throwMsgHelper(itMsg: string, fn: Promise<any>, expectedMsg: string): void;
